+++ app/src/main/java/com/boostcamp/and03/ui/screen/booklist/BooklistUiState.kt
@@ -1,4 +1,6 @@
-package com.boostcamp.and03.ui.screen.booklist.model
+package com.boostcamp.and03.ui.screen.booklist
+
+import com.boostcamp.and03.ui.screen.booklist.model.BookUiModel
 
 data class BooklistUiState(
     val books: List<BookUiModel> = emptyList(),
